requirements.txt
anyio==3.6.2
bcrypt==4.0.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
fonttools==4.39.4
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.12
passlib==1.7.4
pdfkit==1.0.0
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pydyf==0.6.0
pyphen==0.14.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tinycss2==1.2.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
zopfli==0.2.2
